wip: Generate screenshots on each PR.
It is up to the PR author and reviewer(s) to decide whether any changed screenshots match expectations (and the PR can be approved) or represent a regression, and the PR needs changes to either undo the screenshot changes, or to make the screenshots match expectations.

To achieve this, add a new workflow, update-screenshots.yml. This launches Playwright to run tests in tests/, which are expected to take screenshots of different parts of the site.

The screenshots are written to .screenshots in this repository. If there are any changes to those files the changes are made as a new commit on the PR.

The PR author and reviewers can then use GitHub's built-in support for viewing image diffs (onion skin / side-by-side / swipe to reveal, etc) to review the images and make a decision about what to do next.

The PR author is not expected to generate new screenshots as part of preparing their PR. They might have slightly different font versions installed, or different browser versions, all of which can make any screenshot changes they include different from the ones generated on CI. Therefore the screenshots generated on CI are considered "golden".